resolve circular dependency issue w/ abcutil
