+++ AbcUtil.h
@@ -19,8 +19,6 @@
 
 using namespace std;
 
-class Parameter;
-
 namespace ABC {
 
   struct LinearFit {
@@ -150,7 +148,6 @@ namespace ABC {
   int gsl_rng_nonuniform_int(std::vector<double>& weights, const gsl_rng* rng);
 
   double rand_trunc_normal(double mu, double sigma_squared, double min, double max, const gsl_rng* rng);
-  Row rand_trunc_mv_normal(const vector<Parameter*> _model_pars, gsl_vector* mu, gsl_matrix* L, const gsl_rng* rng);
 
   LinearFit* lin_reg(const std::vector<double> &x, const std::vector<double> &y);
